[TKT-15] refactor: 티켓 도메인 서비스 패키지 변경
